+++ mite_data/main.py
@@ -162,6 +162,7 @@ class MibigManager(BaseModel):
     @model_validator(mode="after")
     def download_mibig(self):
         if self.mibig.exists():
+            logger.warning("'mibig' directory exists - SKIP")
             return self
         self.run()
         return self
@@ -286,7 +287,8 @@ class RunManager(BaseModel):
         sys.exit(0)
 
 
-if __name__ == "__main__":
+def main() -> None:
+    """Runs main function"""
     mibig = MibigManager()
     manager = RunManager()
 
@@ -294,3 +296,7 @@ if __name__ == "__main__":
         manager.run_file(path=sys.argv[1])
     except IndexError:
         manager.run_data_dir()
+
+
+if __name__ == "__main__":
+    main()