+++ mite_data/modules/fasta_manager.py
@@ -147,10 +147,12 @@ class FastaManager(BaseModel):
     Attributes:
         src: mite entries location
         fasta: fasta file location
+        timeout: a timeout in seconds
     """
 
     src: DirectoryPath = Path(__file__).parent.parent.joinpath("data/")
     fasta: DirectoryPath = Path(__file__).parent.parent.joinpath("fasta/")
+    timeout: float = 3.0
 
     def update_all(self) -> None:
         """Update metadata of all files (overwrite all)"""
@@ -200,13 +202,22 @@ class FastaManager(BaseModel):
         Returns:
             Tuple of path and data for storage
         """
-        handle = Entrez.efetch(
-            db="protein", id=genpept_acc, rettype="fasta", retmode="text"
-        )
-        fasta_data = handle.read().strip()
-        handle.close()
+        url = "https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi"
+        params = {
+            "db": "protein",
+            "id": genpept_acc,
+            "rettype": "fasta",
+            "retmode": "text",
+        }
+        try:
+            response = requests.get(url, params, timeout=self.timeout)
+        except requests.exceptions.ConnectTimeout as e:
+            raise RuntimeError("Warning: could not connect to NCBI: Timeout") from e
 
-        lines = fasta_data.splitlines()
+        if response is None:
+            raise RuntimeError(f"{mite_acc}: NCBI download failed on ID {genpept_acc}")
+        lines = response.text.strip().splitlines()
+
         if not lines or len(lines) == 1:
             raise ValueError(
                 f"{mite_acc}: No sequence found for GenBank Accession {genpept_acc}"
@@ -237,10 +248,15 @@ class FastaManager(BaseModel):
 
         response = None
         for url in urls:
-            r = requests.get(url)
-            if r.status_code == 200:
-                response = r
-                break
+            try:
+                r = requests.get(url, timeout=self.timeout)
+                if r.status_code == 200:
+                    response = r
+                    break
+            except requests.exceptions.ConnectTimeout as e:
+                raise RuntimeError(
+                    "Warning: could not connect to UniProt: Timeout"
+                ) from e
 
         if response is None:
             raise RuntimeError(